[BUGFIX] Endtime filter not matching index values
The endtime filter was hardcoded to 1970-01-01T01:00:00Z (GMT+1).
However, if the server timezone was a different one, a different value was
generated for unixtime 0. Thus the filter did not match and basically would
exclude all documents as none matches the filter.

Solution is to always generate the endtime filter dynamically.

Fixes: #44276

Change-Id: I62781f0467a62e4a88974ae60a98a55b34c0f9bb