[BUGFIX] Don't add mount pages to index queue with no_search flag
To prevent mount pages added twice to the search index (one without indexing
properties - the original page - and one with indexing properties - the mounted page)
the pages field "no_search" should be checked when adding mount pages to the search
index queue.
Having both pages in the index queue leads to the problem, that, if the mounted page
is indexed after the original one, the solr doc gets overwritten with the data from the
mounted page, because the doc id is the same.
This again leads to problems with e.g. the url (has obsolet mount get parameter) or
treeLevel conditions in the index queue configuration.

Resolves: #47597
Change-Id: Icfb2611c417361bb57ec47a7b4553120e209bd11
Reviewed-on: https://review.typo3.org/22049
Reviewed-by: Stefan Neufeind
Reviewed-by: Ingo Renner
Tested-by: Ingo Renner