[BUGFIX] Remove unusable unicode characters and "�" character
When indexing files it often happens that a file uses another font
than being used on the website. Each font can have different glyphs
in the so called "Private Use Area" of the Unicode standard [1].
When different fonts (TTF, OpenFont) come into play by indexing a
PDF document and displaying it on the webpage it is quite possible
that a glyph in this areas is mapped to a different glyph (or none
at all) in the other font. Usually those special glyphs are used for
displaying bullets, dots, arrows and other special characters in PDF
documents. Instead of adding them to the solr document filter them
away before indexing.
Additionally also the unicode replacement character "�" (often a
black diamond with a white question mark) will get filtered away.
When some conversion process results in a non-valid UTF-8 character
this character is used instead.
[1] http://en.wikipedia.org/wiki/Private_Use_Area_%28Unicode_block%29

Resolves: #54536
Releases: 3.0