[BUGFIX] Fix order of FE initialization in suggest eid script
The suggest script is initializing the frontend itself, and after
changes in 6.x and the way TCA is included and cached, the TYPO3 bootstrap
sequence is changed.

The script is initializing the $TSFE->tmpl->rootline itself, but does so
before the TCA is loaded. This means that if the rootline is not present in
the rootline_cache, relation fields like  media (which is a sys_file_reference
relation) is not correctly resolved as such.

In itself, this does not do much harm, since nothing in the eID script
is dependent on the rootline being calculated correctly, but since this
errornous rootline is then cached, all subsequent requests to that
page will be using the faulty rootline.
If the site uses typoscript levelmedia to render images from the media
field (possibly with slide), this will result in wrong images, or an exception.

Change-Id: Ibda9fea7d013c7ed227c9a8f0d431396fffa636e
Resolves: #55291
Reviewed-on: https://review.typo3.org/27032
Reviewed-by: Ingo Renner
Tested-by: Ingo Renner