[TASK] Ensure type of Dynamic fields
Solr version rejects a document if the value for a field does
not match the field's data type. As TypoScript is solely
"text-based" an _intS field for example may be filled with
an empty string which then prevents the whole document
from being indexed.

This changeset adds a basic type-converter for dynamic
fields from TypoScript.

Releases: 3.0