[BUGFIX] Backend Module calls addFlashMessage which does not work in < 6.2
In the OverviewModuleController $this->addFlashMessage() is called. This
function was implemented in 6.2. That means that this backend module
does not work in TYPO3 6.0 and 6.1.

This patch creates a backport of this function in the AbstractModuleController
that calls the parent function if TYPO3 Version is >= 6.2 and the backported
implementation if TYPO3 Version is < 6.2

Change-Id: I73320d086c8da9c00c9e123ae55d06a0673a578e
Resolves: #60124
Reviewed-on: https://review.typo3.org/31673
Reviewed-by: Phuong Doan <phuong.doan@dkd.de>
Tested-by: Phuong Doan <phuong.doan@dkd.de>
Reviewed-by: Sascha Egerer <sascha@sascha-egerer.de>
Tested-by: Sascha Egerer <sascha@sascha-egerer.de>