[BUGFIX] hierarchical facets broken in multi-language setup
AbstractMenuContentObject::filterMenuPages() checks whether we are showing
a translation (language > 0) and whether hidePagesIfNotTranslatedByDefault
is set active (true/1). If that is the case pages that are not marked as
being a language overlay are filtered and thus not shown.

Since the "pages" in the facets hierarchy are not marked as language overlay,
none of them are rendered and the whole facet appears to be empty.

Fixes: #55812

Change-Id: I6be9d345b273d6c4e85a9faa2b2f970423071333