[BUGFIX] Keep item in IQ when setting future start time
Starting point: Existing item in Index Queue and indexed in Solr.
When setting a new, future start time the item correctly gets
removed from the index. Unfortunately the item also gets removed
from the Index Queue. Thus the item will not be re-indexed once
the start time passes.

The item was actually updated correctly, the garbage collector however
then removed it from both index and Index Queue.

Simply changing the order of execution fixes the issue:
First collect garbage as the item is no longer visible until the start time passes.
Then, since it is an update to the item, execute the record monitor (re-)adding
the item to the Index Queue with the correct future start time.

Change-Id: I48292255a6c29bae65ead3c206310550f3dede2b