[BUGFIX] Detect indexing configuration change
In a situation with multiple indexing configurations for the same table
it can happen that after changes to a record it matches a different
indexing configuration than before the changes.

The new indexing configuration is now set in the Index Queue.

Known issue: If there are multiple indexing configurations matching the
changed record the first one matching is applied.

Change-Id: I8578d08770cb0f00fda74941ef6425dae1bc8308