[TASK] remove composer.lock
composer.lock is unnecessary in a TYPO3 extension
as the EXT cannot be used standalone and a
composer.lock file is only used by the main composer.json
of the project. Therefore if a certain version of a library
is necessary the specific version should be added to
the composer.json instead.