[TASK] Introduce main configuration object
This patch streamlines the typoscript configuration handling.
At various places the typoscript configuration path plugin.tx_solr. is
used, also in context of a content object. So typoscript settings that
should affect a single content element path like plugin.tx_solr_PiResults_Results.
was ignored. Also the flexform settings are merged in the main
configuration object now.

For now it does not support nested solr plugins.

In addition the tests are extended with the scope where the configuration is used.

For backwards compatibility the TypoScriptConfiguration object support array access to read
the TypoScript configuration. In the logrun whe should replace the array access with
speaking methods or "getObjectByPath" or "getValueByPath" calls.

Resolves: #163