[TASK] absRefPrefix not working in cli mode (#492) (#496)
The reason why absRefPrefix is not working in the cli context is, because the webroot is not available in the cli context.

To resolve this. We asume that the webroot is PATH_typo3 (what is most likely the case). If you need to use a different webroot, you can pass it in the indexing scheduler task. When you want to configure a relativ path here, you can use the marker ###PATH_site### and ###PATH_typo3### in the path and they will be replaced with the value of the corresponding constants.

Fixes: #491