[FEATURE] Add hook to manupilate data after they are fetch from database (#609)
In some cases you want to index external records for example
products from a ecommerce platform. In this case the records
are not in the TYPO3 database and we need only the id to index
them with our own indexer. Use this hook to add the needed data
to the tableRecords array.