[TASK] Add cObject support for solr settings
This feature allows you use cObjects in the solr typoscript configuration. This enables you to read the configuration
e.g. from the AdditionConfiguration.php, which is most likely the place to put environment specific settings in deployment scenarios.

The following example shows the usage:

Addition to AdditionalConfiguration.php:

```
   $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['solr']['host'] = 'mysolrserver.de';
```

Usage in typoscript:

```
    plugin.tx_solr.solr {
       host = TEXT
       host {
         value = localhost
         override.data = global:TYPO3_CONF_VARS|EXTCONF|solr|host
       }
    }
```

Fixes: #868