[TASK] Add integration test for GarbageCollector
This PR adds integration tests for the following use cases of the garbage collector:

* content element is deleted (page should be reindexed)
* content element is set to hidden (page should be reindex)
* content element start time is set to date in future (page should be reindex)
* content element end time is set to date in past (page should be reindex)
* page is deleted (page should be removed from the index and index queue)
* page is set to hidden (page should be removed from the index and index queue)

During the tests have been added a bug was fixed, related to the start end endtime usecase.
The check if an item for this record is in the queue does not work correctly for tt_content and pages_language_overlay items since for
the the page with the pid should be checked.

Fixes: #920