[TASK] RecordMonitor does not track records outside the siteroot.
With the configuration "additionalPageIds" it was possible to add additional page ids to an indexing configuration.
When these external records have been updated in the backend by now the RecordMonitor does not recognize these updates.

This PR allows the RecordMonitor to determine all responsible site roots and add the entry to the queue when the queue get's intialized.

Fixes: #485