[TASK] Allow to use closest typoscript template to retrieve configuration
By now, when the configuration is fetched the real pageId is used to retrieve it.
The drawback of this approach is, that in many systems this is an overhead,
because the configuration is only set on the rootlevel or some subpages
(except when e.g. conditions are used).

This pull request introduces an ext_conf_template setting that forces ext_solr
to use the closest page id with a typoscript configuration to retrieve the
configuration. The result is, that the configuration cache has more hits and
the performance in the backend is better

Fixes: #937