[!!!][TASK] Move FilterEncoder and FacetBuilder logic to Domain/Search/ResultSet/Facets package and remove FacetRendererFactory
Implementing new facets types is complicated now, because the code is spreaded in several parts of the extension.
In the long run it would be good to have all logic for a specific facet type(Domain objects, solr query building, urldecoding) in one package.

This PR:

* Moves the logic of facet url decoding to each facet package in Domain/Search/ResultSet/Facets
* Moves the logic of solr facet query url generation to each facat package in Domain/Search/ResultSet/Facets
* Introduces a *Package class that describes all required classes for a new facet type
* Moves the logic of FacetParserRegistry to FacetRegistry. The facet registry now allows you to register facet packages. From the
package you can access the Parser,UrlDecoder and QueryBuilder

Migration:

Changes are only required when you've previously used FacetParserRegistry to register a custom parser. In this case you need to:

* Add a *Package class to you custom facet that extends AbstractFacetPackage
* Register the package with FacetRegistry->registerPackage('PackageClassName', 'customType')
* You can use the default facets types from EXT:solr as example.

Fixes: #1311