[TASK] Make sure that UnitTest do not depend on proper logger configuration
During the investigation for #1469 i recognized that the unit tests depend on a proper logger configuration. This is because the SolrLogManager is instanciated in many places and directly evaluatest the configuration.

To make the unit tests in dependent it should be possible to pass the SolrLogManager from outside and mock it.

Allong with that we should make sure that the logger is only instanciated when really something is logged.

Fixes: #1489