[TASK] Avoid error on update
This PR:

During the update from 6.1.x the localconf.php as a cached state that uses removed code. To avoid this we readd only the signature of those methods to avoid an php error.
The following signatures have been readded (and will be removed after the release of EXT:solr 7.0.0):

* AdministrationModuleManager::registerModule
* CommandResolver::registerPluginCommand
* PluginCommand
* FacetRendererFactory::registerFacetType

Allong with that it makes sence to clear the cache in your TYPO3 project after the autoloader dump:

```
"scripts": {
    "post-autoload-dump": [
      "rm -Rf {$web-dir}/typo3temp/Cache/*"
    ]
  },
```

Fixes: #1518