[TASK] Do not ping solr server before search
Currently the solr server is pinged before every request. In the most cases the solr server is available
and the unavailablility is an exception. Therefore we do not ping the solr server and catch a missing
connection in the catch block. By using this approach we need to do only one request to the solr server in the most
cases instead of two before.

Fixes: #1404