[BUGFIX] Always return value as expected (#1748)
If a field is an array, an array should be returned within Solr
document. This is necessary for e.g. be able to iterate over a property
in a Fluid file. If returning either an array or a string, an exception
in Fluid is thrown because it expects an array.