[TASK] Use solarium to build queries for Solr
In EXT:solr 9.0.0 we want to use the solarium php api instead of the old SolrPhpClient

Since this is a huge task, we splitted it into multiple parts. In this part we change EXT:solr
in order to use as much from solarium as we can to build the Solr queries.

In a later issue/pr we will also use the solarium client, to perform the http requests to solr
and along with that drop the SolrPhpClient.

This pr contains several changes that require changes on you code when you use the EXT:solr api:

* The ParameterBuilder classes (Domain\Search\Query\ParameterBuilder), no get the QueryBuilder passed in the build() method, instead of the Query itself.
* The Query class now inherits from Solarium\QueryType\Select\Query\Query and the QueryBuilder and all ParameterBuilders need to initialize the Query object with the solarium api.
* In non composer mode the solarium api and it's dependecies get loaded from Resources\Private\Php\ComposerLibraries, in composer mode solarium is installed by composer.

Fixes: #2068