[BUGFIX] prevent CURLE_BAD_CONTENT_ENCODING (#1895) (#1902)
When using an nginx server and libcurl 7.57.0 indexing did not work with
libcurl error 61. This was because Content-Encoding "gzip" was requested
but "normal" was returned. To prevent this I changed the header to send
a Content-Type instead of Content-Encoding. This should also work with
apache and also behind Varnish, but should be tested.