[BUGFIX] Do not render id and L param for form for sites managed with site management
When the pageId and the languageId is encoded in the path segment of the url e.g. "/en/search", there is no need,
to keep the in the form as hidden arguments. Nevertheless we need to make sure that they are added when no speaking urls are used (e.g. in a 8 LTS default environment).

This ps:

* Add's variable(addPageAndLanguageId) to the form that indicates if the L and id parameter are required as hidden fields
* The variable is set to false when we have a site in TYPO3 9 LTS managed with the site management because then these arguments are encoded in the path segment of the url

Fixes: #2217