[BUGFIX] Filters in autosuggest get added twice
The filters in the autosuggest currently get added twice. Once from TypoScript and once
from the passed additionalFilters (That originally also come from the TypoScript of the previous request).

Since we can access the TypoScript from the suggest controller as well we should not pass those filters as additionalFilters.

Nevertheless we should keep the evaluation of additionalFilters to allow additionalFilters on a view base.

Related: #2211