[TASK] Prepare different url generation when site management is active
When the site management in TYPO3 9 is used, the language and pageId parameter are part of the encoded url.
In previous versions of EXT:solr we build the frontend url by our own by adding the id, L and MP parameter.

This pr:

* Refactors the url generation by extracting the code into strategy classes
* Implements a SolrSiteStrategy that is able to build the indexing uri as before with index.php?id=xx&L=yy
* Implements a TYPO3SiteStrategy that uses the router to build the indexing uri
* Prepares a different indexing url handling with site management

**Important**: By now the old urls with id and L parameter are used also when site management is active since the
urls that are generated by the router could produce urls that lead to a 404 error.

Fixes: #2234