[BUGFIX] Correctly replace non breaking space entity
In the process of extracting content from HTML, the &nbsp;
entity is being decoded into an UTF-8 char, which in the
end is destroyed with non unicode aware preg_replace calls
which finally leads to json_encode to fail due to encoding errors
and the page indexing fails.

To fix this, we remove unecessary decoding of entities early in the process
so that we later can cleanly remove &nbsp; with a space.