Replace URL with new filter URL (#2557)
If a user clicks on a search result the appropriate page/content is loaded. When going back (browser back) the search page is loaded, but without the chosen filter options. By replacing the current browser URL is is possible to go back to the search page with the chosen filter options already set.