[BUGFIX] Re-enable Integration Tests for TYPO3 v10
Some tests within SearchControllerTests were skipped for TYPO3 v10,
which are re-enabled now.

The main reason why this happened was/is the changed behaviour
of DependencyInjection in TYPO3 v10. This way, a Singleton (in this case
EnvironmentService of Extbase) cannot be replaced during a Functional
Call anymore - which should not be done in any case.

Some frontend-related tests failed for this reason, because it was not possible
to mock the behaviour of TYPO3_MODE=FE in Tests.

Enabling these tests showed that the tests failed as the ViewHelpers did not
receive the current request by the ControllerContext, which is now
optionally injected into all ViewHelpers that generated links.

To run these Integration Tests, a new PHPunit configuration
called "IntegrationFrontendTests.xml" is added which is executed with
the `TYPO3_MODE=FE` constant activated in the configuration.

All other tests continue to work as before.

In addition, some cleanups to avoid deprecation logs in TSFETestBootstrapper
and to execute frontend calls within another frontend calls have been
made including the handling of TYPO3_REQUEST.