[DOCS] Add info to Settings.cfg
This will add the information that is necessary for the docs generator
to add an "Edit on GitHub" button and a link to the repository in the
footer.

The "Edit on GitHub" button will make it easier to contribute to the
docs by creating a change to the docs using only a webbrowser and a
GitHub account.