Prevent duplicate urls for page 0
Reset page argument if set to 0.
As 0 is the default, there is no need to have two different states to represent this.
By resetting, the array will not return the info and links to page 0 will not include the page argument.
This prevents two separate urls to the same source.