[TASK] Introduce Solarium 6
Upgrade from Solarium version 4 to version 6.

Important: Deprecate usage of path with prepending '/solr/' to the
    core name. In EXT:solr >= 11.2 the core name must be used and path be empty.

Introduce use of PSR-18 adapter with the TYPO3 HTTP Client.
The option 'timeout' is not in use anymore and dropped.

Impact: The setting for HTTP client $GLOBALS['TYPO3_CONF_VARS']['HTTP']
   are taken into account again.

Various classes and data type changed according to changes with Solarium.

Note: Class 'Node' will be replaced with 'Endpoint' (solarium) in future.
   Both classes containing mostly the same information.

Fixes: #2636
Fixes: #2549