[BUGFIX:BACKPORT:10] Fix handling of case sensitive variant ids
The VariantsProcessor expects the content of the variant field to be
in lowercase and explicitly applies an mb_strtolower. Since Solr returns
the contents in the expanded field unchanged and the contents might be
case sensitive, the variants may not be recognized correctly and cannot
be accessed.

This issue is fixed by removing the unnecessary mb_strtolower.

Fixes: #2865