[BUGFIX:BACKPORT:11] Fix missing variant field value
If variants are enabled and a returned Solr document has no variants,
the variant field value is not set. Currently variantFieldValue is only
set if variantId is set in expanded, but if there is only one variant in
the Solr index there will be no data in the expanded array.

This comit fixes this issue by ensuring that the variant field value is
set even if there is no data in the expanded array.

Resolves: #2878