[TASK] Remove deprecated cache initialization
Remove deprecated cache initialization from catch block.

The exception written to the log system.
Instead of an new created cache class the method will return null.
This is a different behaviour than before. Consider that earlier an old
TYPO3 structure was in use which support was dropped at least with TYPO3 7.

Fixes: #2782