[TASK] Events, Middleware and Enhancer
Simplify amount of events that are fired for routing and URIs.

Replace own method to find sites with TYPO3 own SiteMatcher.

Changes logical checks according to results of the site matcher.

Removes methods that are not in use anymore.

Adjusts unit tests according to the internal changes.

Introduce new interface in order to test if an enhancer is valid for solr.

Change the name of the enhancer from 'CombinedFacetEnhancer' to
'SolrFacetMaskAndCombineEnhancer' in order to keep clear what
this enhancer does.

Fixes: #2257