[BUGFIX] Fix expected variant results
With Apache Solr 8.8.0 treatment of null and/or 0 values is fixed, see
SOLR-15047. This issue exists since the variant feature was implemented,
we have to adapt our integration tests to respect the new behaviour of
Solr.

This commit adapts the integration tests to expect the correct number of
variant groups.