[BUGFIX:BP:11] remove escaping on suggestion prefix
Since solarium escapes the special uri characters with RFC1738,
there are no needs to do the same within the EXT:Solr.

see:
https://github.com/solariumphp/solarium/blob/97a9df0c2d9ba35db3c747b30c0809702296f18b/src/Component/RequestBuilder/RequestParamsTrait.php#L153

This is the first step of removing the escaping/encoding from EXT:Solr API.
Other components of EXT:solr will be fixed as well in different pull requests.

Fixes: #2917