[BUGFIX] Record indexing doesn't work anymore if page queue is disabled
The initializer MUST ignore only those pages, that match the `additionalWhereClause` directly, and NOT the matched page incl. whole sub-tree of that page.
The Record-Monitoring stack ignores the state of parents-tree,
which match the `additionalWhereClause` and adds the pages to the index queue after first change anyway.
This change fixes this issue and makes it possible to index the pages and records
stored inside of inaccessible sub-tree.

Fixes: #2241
Closes: #2898