[BUGFIX] Changes on sub-tree of mounted source pages are not recognized
This change fixes wrong behavior by changes on sub-tree of mounted pages.
The RecordMonitor has ignored the mounted pages on following scenarios:
* The mounted page is outside of any site/root-page
* The root page is not configured

The variables are renamed to match the right terms to do not mix up the
things like mount point with mount pages, mounted pages and so on
by reading the code.