[TASK] Ensure valid cache ids
Solr extensions are hooking into the DataHandler observing record
changes and partly these hooks are using the TwoLevelCache and the
DataHandler record ids to cache already processed elements.

As the NEW ids processed by the DataHandler can be set by extensions
and may be invalid for the usage with the cache frontend, we should
sanitize the ids to avoid exceptions. This is now ensured by adding
a sanitize method to the TwoLevelCache.

Resolves: #3031