!!![TASK] Improve and Fix TSFE Initialization
The initialization of TSFE within indexing and  Backens modules contexts is refactored.

In this change the setting and usage of $GLOBALS['TSFE'] is removed and replaced by TYPO3s Core Context API.
The "Context" is always cloned instead of using its singleton instance.
The "Context", "Language", "TSFE" and "ServerRequest", which are required for TypoScript parsing in BE-modules and indexing contexts, are highly isolated/capsuled and not visible anymore for all things not belonging to EXT:solr internals.

Byside of isolation/capsulation of TSFE, the language handling is restored 
to pure and default EXT:solr "fallback" mode approach. 
So all page records, which are involved in indexing(All page types and above all all with sys_template or records-to-index), must be translated. Otherwise the translation records will be indexed in default language.

Note: Since TYPO3 11 LTS does not allow to instanciate TSFE for sys folders and spacer, the initialization of TSFE will be done for first and closest page(not spacer or folder) within the site rootline.  

Note: This change is not a final state of EXT:solr 11.5.x 

Fixes: #3092, 3065
Relates: #2976, #2977