[TASK] Refactor IntegrationTest base class : auto import root pages
With this change, the root pages of configured sites are imported by default, if `writeDefaultSolrTestSiteConfiguration()` is called.

To skip the import of records for root pages, the test class property  $skipImportRootPagesAndTemplatesForConfiguredSites must be set to false.

To add or override TypoScript setting please use following typo3/testing-framework methods:
* `addTypoScriptToTemplateRecord()`
* `setUpFrontendRootPage()`

The usage of this feature will be implemented in subsequent commits. 

Fixes: #3092
Relates: #2976, #2977