[FEATURE] Get "free content mode" working
In previous releases of EXT:solr the language handling for "free-content-mode" inconsistent. 
The behavior of "free-content-mode" related records varied in RecordMonitor, Initializing and Indexing contexts, 
which was the source of troubles for mixing overlay records in wrong cores/languages.

This change brings the RecordMonitor, Initializing and Indexing contexts for "free-content-mode" related records 
into the same line, so the "free-content-mode" records are processed the same way.

## About "free-content-mode"-records from TYPO3 docs:
> "Fall back to other language, if the page does not exist in the requested language.
> But always fetch only records of this specific (available) language."
> "TYPO3 will directly fetch the translated records and not care about the records in the default language at all."



Fixes: #2788
Ralates: #2504,