[BUGFIX] Fix autosuggest with non-ascii terms
The introduced ASCII folding filters or language depending normalization
filters lead to issue with the auto suggest function due to the
differing stemming behaviour.

To fix this issue the original token is preserved if possible, this e.g.
allows suggestions for search terms with and without accents. As this
extension might lead to unwanted duplicates a new field textSpellExact
is introduced, which considers non-ascii characters as given.

Ports: #3117
Resolves: #3096