[P:11.5:FEATURE] Improve data update handling
To keep the Solr index up-to-date the indexed records have to be
observed, to minimize the delay and update the index as fast as possible
updates are ensured via DataHandler hooks.

But as intensive monitoring especially in larger websites may slow down
the TYPO3 backend, a delayed processing is introduced. By default the
monitoring is unchanged, but you can switch to a delayed processsing
of the data updates in the extension manager.

Due to the extended data update handling the unit and integration tests
have to be adjusted and extended, this commit contains the required
adjustments.

Resolves: #3153