11.2.0 - Feature Release
We are happy to release EXT:solr 11.2.0.
The focus of this release has been on supporting the latest Apache Solr
version 8.11.1 and on optimizing the data update monitoring.

New in this release
- Apache Solr 8.11.1 support
- Improved data update monitoring and handling

For more information please read the release notes:
https://github.com/TYPO3-Solr/ext-solr/releases/tag/11.2.0

---

How to Get Involved

There are many ways to get involved with Apache Solr for TYPO3:

- Submit bug reports and feature requests on GitHub
- Ask or help or answer questions in our Slack channel
- Provide patches through pull requests or review and comment on
existing pull requests
- Go to www.typo3-solr.com or call dkd to sponsor the ongoing
development of Apache Solr for TYPO3

Support us by becoming an EB partner:
https://shop.dkd.de/Produkte/Apache-Solr-fuer-TYPO3/

or call:
+49 (0)69 - 2475218 0