[BUGFIX] Fix GroupItemPaginate ViewHelper
This change avoids using the "old" widget
syntax for the s:groupItemPaginate" widget
ViewHelper and uses the ResultsPaginator.

This is especially important when using
EXT:solrfluidgrouping.

This change is not complete, as old code
is still flying around and templates
need to be adapted. No tests have been
made but this change pretty much works
for our use-case.