[DOCS] Align README.md with other extensions
Reduce README.md to an abstract and links to important pages
that cover the user's next steps, especially the full
documentation and TER page for installation, but also add-ons
and the paid services.

The details of the contribution are already in CONTRIBUTION.md
and in the full documentation - and obvious if you look at the
git history.

The contact details should be moved to the full documentation.

Keep outgoing links to a minimum to not distract the user.