[TASK] Allow SearchResultSetService instantiation via makeInstance
Extbase ObjectManager is deprecated and should be replaced, but as
EXT:solr doesn't support DI yet and makeInstance doesn't support a
mixture of manual set parameters and DI, the instantiation of
the local ObjectManager of SearchResultSetService is changed and now
also done in the constructor.

With this change SearchResultSetService can now be instantiated via
makeInsstance, even DI is not yet configured.