[RELEASE] Release 11.5.0-rc-2
Second release candidate of 11.5.0 for TYPO3 11 LTS

* TYPO3 11 LTS compatibility
* Improved data update handling

As long as EXT:solr doesn't support DI and the ObjectManager is still
required, the ObjectManager is set/instantiated in the constructor,
preventing initialization issues.

Extbase ObjectManager is deprecated and should be replaced, but as
EXT:solr doesn't support DI yet and makeInstance doesn't support a
mixture of manual set parameters and DI, the instantiation of the local
ObjectManager of SearchResultSetService is changed and now
also done in the constructor.  With this change SearchResultSetService
can now be instantiated via makeInsstance, even DI is not yet
configured.

Prevent type errors by ensuring the right types is used.

Reduce README.md to an abstract and links to important pages that cover
the user's next steps, especially the full
documentation and TER page for installation, but also add-ons and the
paid services.

method (#3098)

Fix exception which happens with PHP8 because the returned page id is
actually a string but an integer is required

Not all kind of errors while initializing the index queue are currently
handled and so an error page or debug trace might be
shown. This commit adds an additional check and implements a more
pleasant error message.

* #3166

* #3160
* #3169
* #3220

This release candidate is not available in TYPO3 TER, if you want to try
this
release, please download and install this release manually from:

* https://github.com/TYPO3-Solr/ext-solr/releases/tag/11.5.0-rc-2

---

There are many ways to get involved with Apache Solr for TYPO3:

* Submit bug reports and feature requests on GitHub
* Ask or help or answer questions in our Slack channel
* Provide patches through pull requests or review and comment on
   existing pull requests
* Go to www.typo3-solr.com or call dkd to sponsor the ongoing
   development of Apache Solr for TYPO3

Support us by becoming an EB partner:
https://shop.dkd.de/Produkte/Apache-Solr-fuer-TYPO3/

or call:
+49 (0)69 - 2475218 0