[BUGFIX] Avoid 404 handler deadlock in special case
The error occurs in a special case when:
- there is no "page" object in TypoScript
- a 404 handler is configured in site configuration
- there is a language without translated root page

In this case, the 404 handler will be stuck in a deadlock, increasing
the RAM usage until it uses all of the available RAM.

This deadlock happens, when the tsfeCache is null. Therefore we ensure
TSFE is set in cache when internal server errors occurred.

Resolves: #3274