[TASK:BP:11.5] Require TYPO3 11.5.14
With EXT:solr 11.5 we're increasingly using local objects, but most
ContentObjects in the TYPO3 core required a global version of the
TypoScriptFrontendController which leads to problems during indexing.
TYPO3 11.5.14 contains some adaptions to support the usage of local
TypoScriptFrontendController objects, these adaptions are fixing this
issue, so we require at least TYPO3 11.5.14.

Resolves: #3239
Resolves: #3244