[TASK] Introduce index queue type setting
Queue types shouldn't be restricted to internal data, which is
already prepared in some classes. As index queue setting 'table'
suggests that a local table is expected, this is replaced with
a new setting 'type'.

queue.[indexConfig].table is now deprecated and will be removed
in v13, 'type' should be used instead.

Resolves: #3365