!!![TASK:BP:11.6] Introduce specific EXT:solr exceptions
Existing exceptions are now based on the recently introduced
generic EXT:solr exception.

Additionally based on the new generic EXT:solr exception a further
exception is introduced, allowing more targeted handling of exceptions:
InvalidIndexQueueInitizalizationPostProcessorException

This is a breaking change as the base excepion changed and the new
exception is no longer based on \UnexpectedValueException.

Ports: #3764
Resolves: #3763